+++ src/main/scala/org/apache/spark/sql/common/DatasketchesFunctionRegistry.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.registrar
+package org.apache.spark.sql.datasketches.common
 
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.FunctionIdentifier
