+++ src/main/scala/org/apache/spark/sql/theta/functions.scala
@@ -15,14 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql
+package org.apache.spark.sql.datasketches.theta
 
+import org.apache.spark.sql.Column
 import org.apache.spark.sql.functions.lit
 
+import org.apache.spark.sql.datasketches.common.DatasketchesScalaFunctionBase
 import org.apache.spark.sql.aggregate.{ThetaSketchAggBuild, ThetaSketchAggUnion}
 import org.apache.spark.sql.expressions.ThetaSketchGetEstimate
+import org.apache.spark.sql.datasketches.common.DatasketchesScalaFunctionBase
 
-object functions_datasketches_theta extends DatasketchesScalaFunctionBase {
+object functions extends DatasketchesScalaFunctionBase {
   def theta_sketch_agg_build(column: Column, lgk: Int): Column = withAggregateFunction {
     new ThetaSketchAggBuild(column.expr, lgk)
   }
